tests: Offload host resolution to the container
So now if the resolution fails the container will crash and try again
the next time it starts. Previously the host was prefetched and
hardcoded into the container so if this failed then it will get stuck
forever. This also got rid of all the sleeps as they were previously
just waiting for DNS to update so the tests overall run faster.